+++ MANIFEST.in
@@ -1,5 +1,5 @@
-include blackjack/rules.json
-include blackjack/schema.json
-include blackjack/strategy_files/basic_hard.csv
-include blackjack/strategy_files/basic_soft.csv
-include blackjack/strategy_files/basic_split.csv
+include py21/rules.json
+include py21/schema.json
+include py21/strategy_files/basic_hard.csv
+include py21/strategy_files/basic_soft.csv
+include py21/strategy_files/basic_split.csv
