--- blackjack/__init__.py
@@ -1,7 +1,0 @@
-from blackjack.card import *
-from blackjack.deck import *
-from blackjack.game import *
-from blackjack.hand import *
-from blackjack.player import *
-import blackjack.strategies
-import blackjack.utils
