--- blackjack/conftest.py
@@ -1,1 +1,0 @@
-# empty file to help pytest discover blackjack package
