+++ py21/conftest.py
@@ -1,0 +1,1 @@
+# empty file to help pytest discover py21 package
