+++ py21/strategies.py
