+++ py21/strategy_files/basic_hard.csv
