+++ py21/strategy_files/basic_soft.csv
